requirements.txt
sanic==23.6.0
redis==5.0.1
dnspython==2.4.2
pymongo==4.6.0
asyncio==3.4.3
aioredis==2.0.1
motor==3.3.1
aiohttp==3.9.1
pyrapidjson==0.5.1
requests==2.31.0
